pyproject.toml
[project]
name = "agent-bar-v2"
version = "0.1.0"
description = "A short project description."
authors = [{ name = "agusramirez", email = "agusramirez@google.com" }]
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.12",
]
keywords = ['agent-catalog', 'adk', 'agent']
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk",
    "requests",
    "google-cloud-aiplatform",
    "pillow", # required by product-ad-generation
    "uvicorn",
    "biopython",
    "google-cloud-storage",
    "google-cloud-bigquery",
    "google-cloud-bigquery-migration",
    "google-cloud-videointelligence",
    "fpdf",
    "google-genai",
    "google-cloud-translate",
    "google-cloud-vision",
    "google-cloud-speech",
    "gcsfs>=2024.11.0",
    "google-cloud-logging>=3.12.0,<4.0.0",
    "google-cloud-aiplatform[evaluation]==1.130.0",    
    "pyarrow",
    "langchain-google-spanner",
    "langchain-google-vertexai",
    "langchain-text-splitters",
    "fastapi>=0.115.8,<1.0.0",
    "asyncpg>=0.30.0,<1.0.0",
    "pygithub",
    "immutabledict",
    "loguru",
    "injector",
    "db-dtypes",
    "pandas",
]

[build-system]
requires = [
  "setuptools >= 40.9.0",
]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["assets*"]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipython>=8.36.0",
    "pytest>=8.3.5",
    "pre-commit>=4.2.0",
    "ruff>=0.11.5",
    "mypy>=1.16.0",
    "deptry>=0.23.0",
    "pytest-cov>=4.0.0",
    "google-adk[eval]",
    "pytest-asyncio",
    "pytest-dotenv",
    "pytest-html",
    "pylint"
]

[tool.agent_packs]
template = "adk"

[tool.mypy]
files = ["src"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true
ignore_missing_imports = true

[tool.basedpyright]
reportAny = false
reportExplicitAny = false
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownVariableType = false
failOnWarnings = false

[tool.pytest.ini_options]
testpaths = ["test"]
python_functions = ["test_*", "eval"]

[tool.ruff]
fix = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["src"]
omit = ["server.py"]

[tool.deptry.per_rule_ignores]
DEP002 = ["uvicorn"]
